cpufreq: Add simple userspace CPU-boost driver
This is a simple CPU-boost driver designed to be controlled by the power HAL's power hints.

Usage: When 1 is written to /sys/module/userspace_cpu_boost/parameters/boost_now by the power HAL, all online CPUs are boosted to a specified boost frequency for a specified duration of time. The default boost frequency is 1134MHz and the default boost duration is 3 seconds.

This is meant to be an open-source solution for the closed-source CPU-boosting behavior that is present on stock AOSP for the Nexus 5.

Signed-off-by: Sultanxda <sultanxda@gmail.com>